* Reimplemented drafts in blog * Content blocks now have an extra column 'extra_id' to identify the extra it's linked with * Fixed random little issues * Fixed typos * Improved code consistency between different modules/code blocks * Improved comments: more explanatory * Merged master & branch 2.1.0, they'll both be the next release

+++ default_www/backend/modules/analytics/actions/delete_landing_page.php
@@ -1,13 +1,12 @@
 <?php
 
 /**
- * BackendAnalyticsDeleteLandingPage
  * This action will delete a landing page
  *
  * @package		backend
  * @subpackage	analytics
  *
- * @author		Annelies Van Extergem <annelies@netlash.com>
+ * @author 		Annelies Van Extergem <annelies@netlash.com>
  * @since		2.0
  */
 class BackendAnalyticsDeleteLandingPage extends BackendBaseActionDelete
@@ -23,21 +22,18 @@ class BackendAnalyticsDeleteLandingPage extends BackendBaseActionDelete
 		$this->id = $this->getParameter('id', 'int');
 
 		// does the item exist
-		if(BackendAnalyticsModel::existsLandingPage($this->id))
+		if($this->id !== null && BackendAnalyticsModel::existsLandingPage($this->id))
 		{
 			// call parent, this will probably add some general CSS/JS or other required files
 			parent::execute();
 
-			// get all data for the user we want to edit
+			// get data
 			$this->record = (array) BackendAnalyticsModel::getLandingPage($this->id);
 
-			// reset some data
-			if(empty($this->record)) $this->redirect(BackendModel::createURLForAction('index') .'&error=non-existing');
-
-			// delete user
+			// delete item
 			BackendAnalyticsModel::deleteLandingPage($this->id);
 
-			// user was deleted, so redirect
+			// item was deleted, so redirect
 			$this->redirect(BackendModel::createURLForAction('index') .'&report=deleted&var='. urlencode($this->record['page_path']));
 		}
